Remove stale react-dom packages from package.lock

+++ src/commonMain/kotlin/sim/rotation/criteria/AbilityCooldownGte.kt
@@ -13,7 +13,7 @@ class AbilityCooldownGte(data: RotationRuleCriterion) : Criterion(Type.ABILITY_C
     }
 
     val seconds: Double? = try {
-        (data.seconds as Int).toDouble().coerceAtLeast(0.0)
+        (data.seconds as Double).coerceAtLeast(0.0)
     } catch (e: NullPointerException) {
         logger.warn { "Field 'seconds' is required for criterion $type" }
         null
